Separating inventoryTest
Separates inventoryTest into chassisInventoryTest and  serverInventoryTest
