+++ Microservices/scripts/SMI/requestdata/chassisInventoryRequestPayload.json
