--- Microservices/scripts/SMI/requestdata/inventoryRequestPayload.json
@@ -1,76 +1,0 @@
-{  
-    "services":{  
-        "hardware":{  
-            "url":"/api/1.0/server/inventory/hardware",
-            "credential":{  
-                "address":"100.68.124.35",
-                "userName":"root",
-                "password":"calvin"
-            }
-        },
-        "software":{  
-            "url":"/api/1.0/server/inventory/software",
-            "credential":{  
-                "address":"100.68.124.35",
-                "userName":"root",
-                "password":"calvin"
-            }
-        },
-        "nics":{  
-            "url":"/api/1.0/server/inventory/nics",
-            "credential":{  
-                "address":"100.68.124.35",
-                "userName":"root",
-                "password":"calvin"
-            }
-        },
-        "bios":{  
-            "url":"/api/1.0/server/inventory/bios",
-            "credential":{  
-                "address":"100.68.124.35",
-                "userName":"root",
-                "password":"calvin"
-            }
-        },
-        "boot":{  
-            "url":"/api/1.0/server/inventory/boot",
-            "credential":{  
-                "address":"100.68.124.35",
-                "userName":"root",
-                "password":"calvin"
-            }
-        },
-        "sel":{  
-            "url":"/api/1.0/server/inventory/logs/get/SEL",
-            "credential":{  
-                "address":"100.68.124.35",
-                "userName":"root",
-                "password":"calvin"
-            }
-        },
-        "lc":{  
-            "url":"/api/1.0/server/inventory/logs/get/LC",
-            "credential":{  
-                "address":"100.68.124.35",
-                "userName":"root",
-                "password":"calvin"
-            }
-        },
-        "chassisDetail":{  
-            "url":"/api/1.0/chassis/inventory/details",
-            "credential":{  
-                "address":"100.68.124.33",
-                "userName":"root",
-                "password":"calvin"
-            }
-        },
-        "chassisSummary":{  
-            "url":"/api/1.0/chassis/inventory/summary",
-            "credential":{  
-                "address":"100.68.124.33",
-                "userName":"root",
-                "password":"calvin"
-            }
-        }
-    }
-}
