+++ Microservices/scripts/SMI/tests/runTests.py
@@ -7,7 +7,7 @@ import unittest
 
 from discoveryTest import DiscoveryMicroserviceTest
 from scpTest import SCPMicroserviceTest
-from inventoryTest import InventoryMicroserviceTest
+from serverInventoryTest import ServerInventoryMicroserviceTest
 from powerThermalTest import PowerThermalMicroserviceTest
 from chassisInventoryTest import ChassisInventoryMicroserviceTest
 
@@ -15,9 +15,9 @@ class microservicesSuite:
         
         discoverySuite = unittest.TestLoader().loadTestsFromTestCase(DiscoveryMicroserviceTest)
         scpSuite = unittest.TestLoader().loadTestsFromTestCase(SCPMicroserviceTest)
-        inventorySuite = unittest.TestLoader().loadTestsFromTestCase(InventoryMicroserviceTest)
+        serverInventorySuite = unittest.TestLoader().loadTestsFromTestCase(ServerInventoryMicroserviceTest)
         powerThermalSuite = unittest.TestLoader().loadTestsFromTestCase(PowerThermalMicroserviceTest) 
         chassisInventorySuite = unittest.TestLoader().loadTestsFromTestCase(ChassisInventoryMicroserviceTest) 
-        allTestsSuite = unittest.TestSuite([discoverySuite, scpSuite, inventorySuite, powerThermalSuite, chassisInventorySuite])
+        allTestsSuite = unittest.TestSuite([discoverySuite, scpSuite, serverInventorySuite, powerThermalSuite, chassisInventorySuite])
 
         unittest.TextTestRunner(verbosity=2).run(allTestsSuite)
